+++ src/utils/EventProvider/misc.ts
@@ -107,4 +107,5 @@ export const onlyGaEvents = [
   'add_payment_info_test',
   'page_load_time',
   'sign_up',
+  'click_card_cashback'
 ];